refactor: renamed inputs
