+++ src/schema/inputs/index.ts
@@ -1,4 +1,4 @@
-export * from './position-attribute';
-export * from './user-profile';
+export * from './position-attribute.input';
+export * from './user-profile.input';
 export * from './user.input';
 export * from './layout.input';
