+++ src/schema/mutation/addPositionAttribute.mutation.ts
@@ -5,7 +5,7 @@ import {
   User,
 } from '../../models';
 import { AppAbility } from '../../security/defineUserAbility';
-import { PositionAttributeInputType } from '../inputs/position-attribute';
+import { PositionAttributeInputType } from '../inputs';
 import { UserType } from '../types';
 
 /**
